+++ src/Schema/Factories/FieldFactory.php
@@ -3,12 +3,13 @@
 namespace Nuwave\Lighthouse\Schema\Factories;
 
 use GraphQL\Language\AST\FieldDefinitionNode;
-use GraphQL\Type\Definition\ResolveInfo;
+use GraphQL\Type\Definition\ResolveInfo as BaseResolveInfo;
 use Illuminate\Container\Container;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Illuminate\Pipeline\Pipeline;
 use Illuminate\Support\Collection;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSetFactory;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\DirectiveLocator;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
@@ -111,10 +112,21 @@ class FieldFactory
             ->thenReturn()
             ->getResolver();
 
-        $fieldValue->setResolver(function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($resolverWithMiddleware) {
-            $resolveInfo->argumentSet = $this->argumentSetFactory->fromResolveInfo($args, $resolveInfo);
+        $fieldValue->setResolver(function ($root, array $args, GraphQLContext $context, BaseResolveInfo $resolveInfo) use ($resolverWithMiddleware) {
+            $wrappedResolveInfo = new ResolveInfo(
+                $resolveInfo->fieldDefinition,
+                $resolveInfo->fieldNodes,
+                $resolveInfo->parentType,
+                $resolveInfo->path,
+                $resolveInfo->schema,
+                $resolveInfo->fragments,
+                $resolveInfo->rootValue,
+                $resolveInfo->operation,
+                $resolveInfo->variableValues,
+                $this->argumentSetFactory->fromResolveInfo($args, $resolveInfo)
+            );
 
-            return $resolverWithMiddleware($root, $args, $context, $resolveInfo);
+            return $resolverWithMiddleware($root, $args, $context, $wrappedResolveInfo);
         });
 
         // To see what is allowed here, look at the validation rules in